Simplify UGen implementation
The audio rate calc function was not being used (since freq is a scalar value),
so it has been removed. The sound is now also based on a sine wave, the
implementation of which is copied from `SinOsc`.
Boid velocity is now deterministically randomised via seed.

Still to do: create complex wave from many frequencies, each affected by a boid.