Attempt to refactor multiple wave data into struct
Note: this still doesn't seem to work for multiple waves.
I have commented out a lot of the data for the wave structs, and they are not
being used anywhere in the algorithm AFAICT.
However - there is one place where I try to set some of the wave struct member
variables to a particular value, and doing so crashes scsynth.
What is even more confusing is that moving this exact loop into the constructor
allows it to run fine.